chore: move constants to primitives
